made it work with Ruby 1.9
Ruby 1.9 has a different treatment of "character" data type.

In 1.8.7, it's an integer. 1.9.2 treats it as one-char string.
Thus, for 1.9.2 to work, it is necessary to insert the conversion
functions ord and chr at appropriate places. 1.8.7 will treat ord
on integers as no-op. chr is always invoked on integers, as is
correct. This will make a slight performance hit, as an additional
function call is performed for each pixel access.